Add way of triggering saving.
This involves calculating the hash of the request as a way of verifying
the user (hopefully) generated it. Note that this has no UI or niceties
around it right now and is merely to ensure functionality. Server URL is
stored in the message as well in case settings change after the request
was sent.